refactor: move ai sse plumbing to windmill-ai
