+++ backend/windmill-ai/Cargo.toml
@@ -21,6 +21,7 @@ windmill-mcp = { workspace = true, optional = true }
 
 async-trait.workspace = true
 base64.workspace = true
+eventsource-stream.workspace = true
 reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -29,6 +30,7 @@ uuid.workspace = true
 lazy_static.workspace = true
 tracing.workspace = true
 tokio.workspace = true
+tokio-stream.workspace = true
 
 # Bedrock (optional)
 aws-config = { workspace = true, optional = true }
