+++ backend/windmill-ai/src/lib.rs
@@ -5,4 +5,6 @@ pub mod ai_google;
 pub mod ai_providers;
 pub mod ai_types;
 pub mod query_builder;
+pub mod sse;
 pub mod types;
+pub mod utils;
