+++ backend/windmill-worker/src/ai/utils.rs
@@ -1,4 +1,6 @@
 pub use crate::ai::types::McpToolSource;
+pub use windmill_ai::utils::{extract_text_content, should_use_structured_output_tool};
+
 use crate::ai::types::ToolDef;
 use anyhow::Context;
 use serde_json::value::RawValue;
@@ -8,7 +10,6 @@ use std::{
     sync::Arc,
 };
 use uuid::Uuid;
-use windmill_ai::ai_providers::AIProvider;
 use windmill_common::flows::FlowModuleValue;
 use windmill_common::{
     db::DB,
@@ -321,11 +322,6 @@ pub fn get_step_name_from_flow(
             .map(|s| s.to_string())
             .unwrap_or_else(|| format!("AI Agent Step {}", flow_step_id)),
     )
-}
-
-/// AWS Bedrock do not handle structured output query param, so we use a tool for structured output. Same for every Claude models.
-pub fn should_use_structured_output_tool(provider: &AIProvider, model: &str) -> bool {
-    model.contains("claude") || provider == &AIProvider::AWSBedrock
 }
 
 /// Cleanup MCP clients by gracefully shutting down connections
@@ -712,22 +708,4 @@ pub fn any_tool_needs_previous_result(tools: &[Tool]) -> bool {
         }
         false
     })
-}
-
-/// Extract text content from OpenAIContent, joining parts with space if multiple
-pub fn extract_text_content(content: &OpenAIContent) -> String {
-    match content {
-        OpenAIContent::Text(text) => text.clone(),
-        OpenAIContent::Parts(parts) => parts
-            .iter()
-            .filter_map(|p| {
-                if let ContentPart::Text { text } = p {
-                    Some(text.as_str())
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<_>>()
-            .join(""),
-    }
 }
